Add base class for cyclic patrolling algorithms
