+++ Assets/Scripts/Algorithms/Patrolling/CyclicAlgorithms/SingleCycleTSP.cs
@@ -23,9 +23,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-using Maes.Algorithms.Patrolling.Components;
 using Maes.Map;
-using Maes.Robot;
 
 using UnityEngine;
 
@@ -33,48 +31,18 @@ namespace Maes.Algorithms.Patrolling
 {
     /// <summary>
     /// THIS ALGORITHMS HAS FACTORIAL TIME-COMPLEXITY OF THE SIZE OF THE ROBOTS PARTITION!
-    /// Implementation of the Single Cycle algorithm but using TSP: https://doi.org/10.1007/978-3-540-28645-5_48.
+    /// Implementation of the Single Cycle algorithm but using exact TSP solver instead christofides as in: https://doi.org/10.1007/978-3-540-28645-5_48.
     /// An implementation can be found here: https://github.com/matteoprata/DRONET-for-Patrolling/blob/main_july_2023/src/patrolling/tsp_cycle.py
     /// </summary>
-    public sealed class SingleCycleTSP : PatrollingAlgorithm
+    public sealed class SingleCycleTSP : BaseCyclicAlgorithm
     {
-        public override string AlgorithmName => "SingleCycle Algorithm";
-
-        // Set by CreateComponents
-        private GoToNextVertexComponent _goToNextVertexComponent = null!;
-        private CollisionRecoveryComponent _collisionRecoveryComponent = null!;
-        private List<Vertex> _patrollingCycle = new();
-
-        protected override IComponent[] CreateComponents(IRobotController controller, PatrollingMap patrollingMap)
-        {
-            _goToNextVertexComponent = new GoToNextVertexComponent(NextVertex, this, controller, patrollingMap);
-            _collisionRecoveryComponent = new CollisionRecoveryComponent(controller, _goToNextVertexComponent);
-
-            return new IComponent[] { _goToNextVertexComponent, _collisionRecoveryComponent };
-        }
-
-        private Vertex NextVertex(Vertex currentVertex)
-        {
-            if (_patrollingCycle.Count == 0)
-            {
-                _patrollingCycle = CreatePatrollingCycle(currentVertex);
-            }
-            return NextVertexInCycle(currentVertex);
-        }
-
-        private Vertex NextVertexInCycle(Vertex currentVertex)
-        {
-            // Get the next vertex in the cycle of _patrollingCycle
-            var currentIndex = _patrollingCycle.IndexOf(currentVertex);
-            var nextIndex = (currentIndex + 1) % _patrollingCycle.Count;
-            return _patrollingCycle[nextIndex];
-        }
+        public override string AlgorithmName => "SingleCycleTSP Algorithm";
 
         /// <summary>
-        /// Use TSP to make the optimal cycle of all vertices in this robots partition
+        /// Use exact TSP solver to make the optimal cycle of all vertices in this robots partition.
         /// </summary>
         /// <param name="startVertex"></param>
-        private List<Vertex> CreatePatrollingCycle(Vertex startVertex)
+        protected override List<Vertex> CreatePatrollingCycle(Vertex startVertex)
         {
             var verticesInPartition = _patrollingMap.Vertices.Where(v => v.Partition == startVertex.Partition).ToList();
             var bestPath = new List<Vertex>();
